@@ -14,7 +14,7 @@ description: A new Flutter project.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.2.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -22,10 +22,30 @@ dependencies:
   firebase_core: ^0.4.0+1
   firebase_auth: ^0.11.1+3
   firebase_database: ^3.0.7
+  cloud_firestore: ^0.12.9+5
   firebase_storage: ^3.0.6
+  uuid: ^2.0.2
+  image: ^2.1.4
+  animator: ^1.0.0+1
+  path_provider: ^1.3.0
+  firebase_messaging: ^5.1.6
+  timeago: ^2.0.20
+  cached_network_image: ^1.1.1
   circular_profile_avatar: ^1.0.2
   json_serializable: ^3.2.2
   geolocator: ^5.1.3
+  flutter_icons: ^0.3.1
+  flutter_svg: ^0.14.2
+  flutter_form_builder: ^3.5.5
+  provider: ^3.1.0
+  line_icons: ^0.2.0
+  rflutter_alert: ^1.0.3
+  image_cropper: ^1.1.0
+  image_picker: ^0.6.1+7
+  flutter_rounded_date_picker: ^0.2.3
+  dots_indicator: ^1.0.0
+
+  # flutter_facebook_login: ^1.1.1
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
@@ -45,9 +65,10 @@ flutter:
   # the material Icons class.
   uses-material-design: true
   # To add assets to your application, add an assets section, like this:
-  # assets:
-  #  - images/a_dot_burr.jpeg
-  #  - images/a_dot_ham.jpeg
+  assets:
+    - assets/icons/facebook.png
+    - assets/images/user-placeholder.png
+    - assets/images/search.svg
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware.
   # For details regarding adding assets from package dependencies, see
@@ -57,17 +78,26 @@ flutter:
   # "family" key with the font family name, and a "fonts" key with a
   # list giving the asset and other descriptors for the font. For
   # example:
-  # fonts:
-  #   - family: Schyler
-  #     fonts:
-  #       - asset: fonts/Schyler-Regular.ttf
-  #       - asset: fonts/Schyler-Italic.ttf
-  #         style: italic
-  #   - family: Trajan Pro
-  #     fonts:
-  #       - asset: fonts/TrajanPro.ttf
-  #       - asset: fonts/TrajanPro_Bold.ttf
-  #         weight: 700
+  fonts:
+    - family: InstrumentIcons
+      fonts:
+        - asset: assets/fonts/InstrumentIcons.ttf
+          weight: 400
+    - family: Montserrat
+      fonts:
+        - asset: assets/fonts/Montserrat/Montserrat-Medium.ttf
+          weight: 500
+        - asset: assets/fonts/Montserrat/Montserrat-Regular.ttf
+          weight: 400
+        - asset: assets/fonts/Montserrat/Montserrat-Bold.ttf
+          weight: 600
+
+    - family: Muli
+      fonts:
+        - asset: assets/fonts/Muli/Muli-Regular.ttf
+          weight: 400
+        - asset: assets/fonts/Muli/Muli-Bold.ttf
+          weight: 600
   #
   # For details regarding fonts from package dependencies,
   # see https://flutter.dev/custom-fonts/#from-packages
