@@ -1,98 +1,104 @@
+import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:jammerz/views/LoginScreen.dart';
-import './ChatScreen.dart';
+import 'package:jammerz/views/OnboardingScreen.dart';
+import 'package:line_icons/line_icons.dart';
+import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:provider/provider.dart';
+import '../presentation/InstrumentIcons.dart';
+import './ChatsScreen.dart';
+import './ProfileScreen.dart';
+import './TimelineScreen.dart';
+import './DiscoverScreen.dart';
+import './ActivityScreen.dart';
+import './UI/Progress.dart';
+import '../Utils.dart';
+import '../models/user.dart';
 
-class HomeScreen extends StatelessWidget {
+User currentUser;
+
+class HomeScreen extends StatefulWidget {
   static const routeName = '/home-screen';
 
   @override
+  _HomeScreenState createState() => _HomeScreenState();
+}
+
+class _HomeScreenState extends State<HomeScreen> {
+  PageController _pageController;
+  int pageIndex = 0;
+
+  @override
+  void initState() {
+    super.initState();
+    _pageController = PageController();
+  }
+
+  @override
+  void dispose() {
+    _pageController.dispose();
+    super.dispose();
+  }
+
+  onPageChanged(int pageIndex) {
+    setState(() {
+      this.pageIndex = pageIndex;
+    });
+  }
+
+  onTap(int pageIndex) {
+    _pageController.animateToPage(pageIndex,
+        duration: Duration(milliseconds: 300), curve: Curves.easeInOut);
+  }
+
+  @override
   Widget build(BuildContext context) {
-    var screenSize = MediaQuery.of(context).size;
-    return Scaffold(
-      floatingActionButtonLocation: FloatingActionButtonLocation.centerDocked,
-      floatingActionButton: FloatingActionButton(
-        child: const Icon(Icons.music_note),
-        onPressed: () {},
-      ),
-      bottomNavigationBar: BottomAppBar(
-        shape: CircularNotchedRectangle(),
-        notchMargin: 4.0,
-        child: new Row(
-          mainAxisSize: MainAxisSize.max,
-          mainAxisAlignment: MainAxisAlignment.spaceBetween,
-          children: <Widget>[
-            IconButton(
-              icon: Icon(Icons.chat_bubble),
-              onPressed: () =>
-                  Navigator.pushNamed(context, ChatScreen.routeName),
-            ),
-            IconButton(
-              icon: Icon(Icons.group),
-              onPressed: () {},
-            ),
-          ],
-        ),
-      ),
-      appBar: AppBar(
-        title: Text("Jammerz"),
-        actions: <Widget>[
-          IconButton(
-            icon: Icon(Icons.settings),
-            onPressed: () {},
-          ),
-          IconButton(
-            icon: Icon(
-              Icons.exit_to_app,
-            ),
-            onPressed: () {
-              Navigator.pushReplacementNamed(context, LoginScreen.routeName);
-            },
-          )
-        ],
-      ),
-      body: Container(
-        height: screenSize.height,
-        width: screenSize.width,
-        child: SingleChildScrollView(
-            child: Column(
-          children: <Widget>[
-            Card(
-              elevation: 5,
-              child: Row(
-                mainAxisAlignment: MainAxisAlignment.start,
+    return FutureBuilder<DocumentSnapshot>(
+      future:
+          Firestore.instance.collection('users').document(Utils.getUid()).get(),
+      builder: (context, AsyncSnapshot<DocumentSnapshot> snapshot) {
+        if (snapshot.error == null && snapshot.data != null) {
+          if (!snapshot.data.exists) {
+            return Scaffold(
+              body: OnboardingScreen(),
+            );
+          } else {
+            currentUser = User.fromDocument(snapshot.data);
+            return Scaffold(
+              body: PageView(
                 children: <Widget>[
-                  Padding(
-                    child: CircleAvatar(
-                      radius: 50,
-                      backgroundImage: NetworkImage(
-                          "https://thumbnailer.mixcloud.com/unsafe/1200x628/tmp/7/4/2/8/fac1-7b75-4a97-a54c-02d8d853fa48"),
-                    ),
-                    padding: EdgeInsets.all(10),
-                  ),
-                  Column(
-                    children: <Widget>[
-                      Padding(
-                        padding:
-                            EdgeInsets.symmetric(vertical: 0, horizontal: 10),
-                        child: Text(
-                          "Omer Yampel",
-                          style: TextStyle(fontWeight: FontWeight.bold),
-                        ),
-                      ),
-                      Padding(
-                        padding: EdgeInsets.all(0),
-                        child: Text(
-                          "Guitarist",
-                        ),
-                      ),
-                    ],
-                  ),
+                  TimelineScreen(),
+                  ChatsScreen(),
+                  DiscoverScreen(),
+                  ActivityScreen(),
+                  ProfileScreen()
+                ],
+                controller: _pageController,
+                onPageChanged: onPageChanged,
+                physics: NeverScrollableScrollPhysics(),
+                pageSnapping: false,
+              ),
+              bottomNavigationBar: CupertinoTabBar(
+                currentIndex: pageIndex,
+                onTap: onTap,
+                activeColor: Theme.of(context).primaryColor,
+                items: [
+                  BottomNavigationBarItem(icon: Icon(LineIcons.fire)),
+                  BottomNavigationBarItem(icon: Icon(LineIcons.comments_o)),
+                  BottomNavigationBarItem(
+                      icon: Icon(
+                    InstrumentIcons.hand,
+                    size: 35.0,
+                  )),
+                  BottomNavigationBarItem(icon: Icon(LineIcons.bell_o)),
+                  BottomNavigationBarItem(icon: Icon(LineIcons.user)),
                 ],
               ),
-            ),
-          ],
-        )),
-      ),
+            );
+          }
+        } else {
+          return circularProgress(context);
+        }
+      },
     );
   }
 }
